Overhauled lighting system to use a single shadow atlas. This will let us support many more lights at once, without using excessive amounts of texture binding. Some refactoring is needed to clean up the code.
