+++ Graphics Engine/Graphics Engine.vcxproj
@@ -192,7 +192,7 @@
     <ClCompile Include="src\rendering\core\MeshBuilder.cpp" />
     <ClCompile Include="src\rendering\core\Camera.cpp" />
     <ClCompile Include="src\rendering\core\Light.cpp" />
-    <ClCompile Include="src\rendering\AssetManager.cpp" />
+    <ClCompile Include="src\rendering\ResourceManager.cpp" />
     <ClCompile Include="src\rendering\core\Texture.cpp" />
     <ClCompile Include="src\rendering\core\TextureBuilder.cpp" />
     <ClCompile Include="src\rendering\imgui\imgui.cpp" />
@@ -203,6 +203,7 @@
     <ClCompile Include="src\rendering\imgui\imgui_tables.cpp" />
     <ClCompile Include="src\rendering\imgui\imgui_widgets.cpp" />
     <ClCompile Include="src\rendering\MarchingCubes.cpp" />
+    <ClCompile Include="src\rendering\TextureManager.cpp" />
     <ClCompile Include="src\rendering\util\OBJFile.cpp" />
     <ClCompile Include="src\rendering\util\PNGFile.cpp" />
     <ClCompile Include="src\rendering\RenderRequest.cpp" />
@@ -263,10 +264,11 @@
     <ClInclude Include="src\rendering\core\MeshBuilder.h" />
     <ClInclude Include="src\rendering\core\Camera.h" />
     <ClInclude Include="src\rendering\core\Light.h" />
-    <ClInclude Include="src\rendering\AssetManager.h" />
+    <ClInclude Include="src\rendering\ResourceManager.h" />
     <ClInclude Include="src\rendering\RenderRequest.h" />
     <ClInclude Include="src\rendering\Shader.h" />
     <ClInclude Include="src\rendering\ShaderManager.h" />
+    <ClInclude Include="src\rendering\TextureManager.h" />
     <ClInclude Include="src\rendering\util\CPUTimer.h" />
     <ClInclude Include="src\rendering\util\GPUTimer.h" />
     <ClInclude Include="src\rendering\util\OBJFile.h" />
