+++ Graphics Engine/imgui.ini
@@ -1,6 +1,7 @@
 [Window][Debug##Default]
 Pos=28,18
 Size=629,522
+Collapsed=1
 
 [Window][Dear ImGui Demo]
 Pos=-2,24
