+++ Graphics Engine/src/rendering/core/Light.h
@@ -4,6 +4,7 @@
 
 #include "rendering/Direct3D11.h"
 #include "rendering/Shader.h"
+#include "rendering/util/TextureAtlas.h"
 
 #include "math/Color.h"
 #include "math/Matrix4.h"
@@ -12,11 +13,15 @@ namespace Engine {
 namespace Graphics {
 
 enum ShadowMapQuality {
-    QUALITY_0 = 64, 
+    QUALITY_0 = 64,
     QUALITY_1 = 128,
     QUALITY_2 = 256,
     QUALITY_3 = 512,
-    QUALITY_4 = 1024
+    QUALITY_DEFAULT = QUALITY_1
+};
+
+struct AtlasTransform {
+    float x, y, width, height;
 };
 
 // LightComponent Class:
@@ -25,23 +30,18 @@ enum ShadowMapQuality {
 // The "direction" of the light's view is given by the direction of its rotated
 // +Z axis. To rotate a light, simply rotate its transform.
 class Light : public Camera {
+  public:
+    // TEMP
+    static TextureAtlas* shadow_atlas;
+
   protected:
     // Light emission color
     Color color;
 
-    // Shadow map texture, and associated data to
-    // let us bind it to the shader and render to it.
-    ID3D11Texture2D* shadow_map;
-    UINT shadowmap_width;
-    UINT shadowmap_height;
-
-    // Enables rendering to the texture
+    // Viewport that must be used to write into the part
+    // of the texture atlas allocated for the light
     D3D11_VIEWPORT viewport;
-    ID3D11DepthStencilView* depth_stencil_view;
-
-    // Enables use of / sampling of the texture in shaders
-    ID3D11ShaderResourceView* shader_resource_view;
-    ID3D11SamplerState* sampler_state;
+    AtlasTransform atlas_transform;
 
   public:
     Light(ID3D11Device* device, ShadowMapQuality quality);
@@ -49,15 +49,8 @@ class Light : public Camera {
 
     // Accessors of the Light's Data
     Color& getColor();
-
     D3D11_VIEWPORT& getViewport();
-    ID3D11DepthStencilView*& getDepthView();
-
-    ID3D11ShaderResourceView*& getShaderView();
-    ID3D11SamplerState*& getSampler();
-
-    UINT getWidth();
-    UINT getHeight();
+    AtlasTransform& getAtlasTransform();
 
     const Matrix4 getProjectionMatrix(void) const;
 };
