+++ Graphics Engine/src/rendering/core/Texture.cpp
@@ -7,16 +7,21 @@
 namespace Engine {
 namespace Graphics {
 
+Texture::Texture(UINT _width, UINT _height) {
+    width = _width;
+    height = _height;
+}
 Texture::~Texture() {
     if (texture != nullptr)
         texture->Release();
-    if (view != nullptr)
-        view->Release();
+    if (shader_view != nullptr)
+        shader_view->Release();
 }
 #if defined(_DEBUG)
-void Texture::displayImGui() {
-    ImGui::Image((ImTextureID) (intptr_t) view,
-                 ImVec2(width, height));
+void Texture::displayImGui() { displayImGui(256); }
+void Texture::displayImGui(float display_width) {
+    ImGui::Image((ImTextureID)(intptr_t)shader_view,
+                 ImVec2(display_width, display_width * height / width));
 }
 #endif
