+++ Graphics Engine/src/rendering/util/TextureAtlas.h
@@ -1,3 +1,4 @@
+
 #pragma once
 
 #include <vector>
@@ -23,12 +24,12 @@ typedef unsigned int UINT;
 // Stores the data regarding where an individual "texture" is
 // located in the atlas. Can be used, with texture coordinates (x,y),
 // x,y in [0,1], to reference the actual texel coordinate in the TextureAtlas.
-struct TextureAllocation {
+struct AtlasAllocation {
     UINT x, y; // Top-left corner in atlas
     UINT width, height;
 
-    TextureAllocation(UINT x, UINT y, UINT w, UINT h);
-    TextureAllocation();
+    AtlasAllocation(UINT x, UINT y, UINT w, UINT h);
+    AtlasAllocation();
 
     UINT area() const;
 };
@@ -41,22 +42,24 @@ struct TextureAllocation {
 // bottom-right (note that Y grows top-down).
 class TextureAtlas {
   private:
-    Texture texture;
+    Texture* texture;
 
-    std::vector<TextureAllocation> allocations;
+    std::vector<AtlasAllocation> allocations;
 
     // Used in the 2D Rectangle Packing (see allocateTexture())
-    std::vector<TextureAllocation> open_regions;
+    std::vector<AtlasAllocation> open_regions;
 
   public:
-    TextureAtlas(UINT width, UINT height);
+    TextureAtlas(Texture* texture);
     ~TextureAtlas();
 
     void initialize();
 
+    const Texture* getTexture() const;
+
     // Use a 2D Rectangle Packing algorithm to pack the
     // allocated textures tightly.
-    const TextureAllocation& getAllocation(UINT index) const;
+    const AtlasAllocation& getAllocation(UINT index) const;
     Vector2 getAtlasCoordinates(UINT texture, Vector2 tex_coords);
 
     UINT allocateTexture(UINT tex_width, UINT tex_height);
